moved create_test_client
