+++ starlite/testing/__init__.py
@@ -1,4 +1,5 @@
+from starlite.testing.create_test_client import create_test_client
 from starlite.testing.request_factory import RequestFactory
-from starlite.testing.test_client import TestClient, create_test_client
+from starlite.testing.test_client import TestClient
 
 __all__ = ("TestClient", "RequestFactory", "create_test_client")
